Remove pool_disjoint_impl.*
Expose disjoint_pool only through pool/pool_disjoint.h
